requirements.txt
Flask>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
gunicorn>=21.0.0
